// src/pages/training/types.d.ts
import { ImageCMS, TextCMS } from "@/types";

export interface TransportContent {
  banner: ImageCMS;
  firstAbout: ImageCMS;
  secondAbout: ImageCMS;
  thirdAbout: ImageCMS;
  map: ImageCMS;
  tabs: ImageCMS[];
  secondTabs: ImageCMS[];
  simpleBanner: ImageCMS;
  logo: ImageCMS;
  differentials: TextCMS;
  simpleInformation: TextCMS;
  otherTrainings: ImageCMS[];
}